fixed nomenclature in DeccelQuartic
